Devolve more processing to RPMC
Helper method delegates are now formed in the RasterPropMonitorComputer.
