+++ RasterPropMonitor/Core/RPMVCEvaluators.cs
@@ -26,428 +26,419 @@ namespace JSI
     public partial class RPMVesselComputer : VesselModule
     {
         // Plugin-modifiable Evaluators
-        //private Func<bool> evaluateMechJebAvailable;
-        private Func<double> evaluateAngleOfAttack;
-        //private Func<double> evaluateDeltaV;
-        //private Func<double> evaluateDeltaVStage;
-        //private Func<double> evaluateDragForce;
-        //private Func<double> evaluateLandingError;
-        //private Func<double> evaluateLandingAltitude;
-        //private Func<double> evaluateLandingLatitude;
-        //private Func<double> evaluateLandingLongitude;
-        //private Func<double> evaluateLiftForce;
-        private Func<double> evaluateSideSlip;
-        private Func<double> evaluateTerminalVelocity;
-        private Func<double> evaluateTimeToImpact;
+        //private Func<double> evaluateAngleOfAttack;
+        //private Func<double> evaluateSideSlip;
+        //private Func<double> evaluateTerminalVelocity;
+        //private Func<double> evaluateTimeToImpact;
 
-        //--- Plugin-enabled evaluators
-        #region PluginEvaluators
-        private float EvaluateAngleOfAttack()
-        {
-            if (evaluateAngleOfAttack == null)
-            {
-                Func<double> accessor = null;
+        ////--- Plugin-enabled evaluators
+        //#region PluginEvaluators
+        //private float EvaluateAngleOfAttack()
+        //{
+        //    if (evaluateAngleOfAttack == null)
+        //    {
+        //        Func<double> accessor = null;
 
-                accessor = (Func<double>)GetInternalMethod("JSIFAR:GetAngleOfAttack", typeof(Func<double>));
-                if (accessor != null)
-                {
-                    double value = accessor();
-                    if (double.IsNaN(value))
-                    {
-                        accessor = null;
-                    }
-                }
+        //        accessor = (Func<double>)GetInternalMethod("JSIFAR:GetAngleOfAttack", typeof(Func<double>));
+        //        if (accessor != null)
+        //        {
+        //            double value = accessor();
+        //            if (double.IsNaN(value))
+        //            {
+        //                accessor = null;
+        //            }
+        //        }
 
-                if (accessor == null)
-                {
-                    evaluateAngleOfAttack = FallbackEvaluateAngleOfAttack;
-                }
-                else
-                {
-                    evaluateAngleOfAttack = accessor;
-                }
-            }
+        //        if (accessor == null)
+        //        {
+        //            evaluateAngleOfAttack = FallbackEvaluateAngleOfAttack;
+        //        }
+        //        else
+        //        {
+        //            evaluateAngleOfAttack = accessor;
+        //        }
+        //    }
 
-            return (float)evaluateAngleOfAttack();
-        }
+        //    return (float)evaluateAngleOfAttack();
+        //}
 
-        private VariableEvaluator AngleOfAttack()
-        {
-            Func<double> accessor = null;
+        //private VariableEvaluator AngleOfAttack()
+        //{
+        //    Func<double> accessor = null;
 
-            accessor = (Func<double>)GetInternalMethod("JSIFAR:GetAngleOfAttack", typeof(Func<double>));
-            if (accessor != null)
-            {
-                double value = accessor();
-                if (double.IsNaN(value))
-                {
-                    accessor = null;
-                }
-            }
+        //    accessor = (Func<double>)GetInternalMethod("JSIFAR:GetAngleOfAttack", typeof(Func<double>));
+        //    if (accessor != null)
+        //    {
+        //        double value = accessor();
+        //        if (double.IsNaN(value))
+        //        {
+        //            accessor = null;
+        //        }
+        //    }
 
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateAngleOfAttack(); };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
-            }
-        }
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateAngleOfAttack(); };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
+        //    }
+        //}
 
-        private VariableEvaluator DeltaV()
-        {
-            Func<double> accessor = null;
+        //private VariableEvaluator DeltaV()
+        //{
+        //    Func<double> accessor = null;
 
-            accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetDeltaV", typeof(Func<double>));
-            if (accessor != null)
-            {
-                double value = accessor();
-                if (double.IsNaN(value))
-                {
-                    accessor = null;
-                }
-            }
+        //    accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetDeltaV", typeof(Func<double>));
+        //    if (accessor != null)
+        //    {
+        //        double value = accessor();
+        //        if (double.IsNaN(value))
+        //        {
+        //            accessor = null;
+        //        }
+        //    }
 
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return (actualAverageIsp * gee) * Math.Log(totalShipWetMass / (totalShipWetMass - resources.PropellantMass(false))); };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
-            }
-        }
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return (actualAverageIsp * gee) * Math.Log(totalShipWetMass / (totalShipWetMass - resources.PropellantMass(false))); };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
+        //    }
+        //}
 
-        private VariableEvaluator DeltaVStage()
-        {
-            Func<double> accessor = null;
+        //private VariableEvaluator DeltaVStage()
+        //{
+        //    Func<double> accessor = null;
 
-            accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetStageDeltaV", typeof(Func<double>));
-            if (accessor != null)
-            {
-                double value = accessor();
-                if (double.IsNaN(value))
-                {
-                    accessor = null;
-                }
-            }
+        //    accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetStageDeltaV", typeof(Func<double>));
+        //    if (accessor != null)
+        //    {
+        //        double value = accessor();
+        //        if (double.IsNaN(value))
+        //        {
+        //            accessor = null;
+        //        }
+        //    }
 
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return (actualAverageIsp * gee) * Math.Log(totalShipWetMass / (totalShipWetMass - resources.PropellantMass(true))); };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
-            }
-        }
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return (actualAverageIsp * gee) * Math.Log(totalShipWetMass / (totalShipWetMass - resources.PropellantMass(true))); };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
+        //    }
+        //}
 
-        private VariableEvaluator DragAccel()
-        {
-            Func<double> accessor = null;
+        //private VariableEvaluator DragAccel()
+        //{
+        //    Func<double> accessor = null;
 
-            accessor = (Func<double>)GetInternalMethod("JSIFAR:GetDragForce", typeof(Func<double>));
-            if (accessor != null)
-            {
-                double value = accessor();
-                if (double.IsNaN(value))
-                {
-                    accessor = null;
-                }
-            }
+        //    accessor = (Func<double>)GetInternalMethod("JSIFAR:GetDragForce", typeof(Func<double>));
+        //    if (accessor != null)
+        //    {
+        //        double value = accessor();
+        //        if (double.IsNaN(value))
+        //        {
+        //            accessor = null;
+        //        }
+        //    }
 
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateDragForce() / totalShipWetMass; };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor() / totalShipWetMass; };
-            }
-        }
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateDragForce() / totalShipWetMass; };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor() / totalShipWetMass; };
+        //    }
+        //}
 
-        private VariableEvaluator DragForce()
-        {
-            Func<double> accessor = null;
+        //private VariableEvaluator DragForce()
+        //{
+        //    Func<double> accessor = null;
 
-            accessor = (Func<double>)GetInternalMethod("JSIFAR:GetDragForce", typeof(Func<double>));
-            if (accessor != null)
-            {
-                double value = accessor();
-                if (double.IsNaN(value))
-                {
-                    accessor = null;
-                }
-            }
+        //    accessor = (Func<double>)GetInternalMethod("JSIFAR:GetDragForce", typeof(Func<double>));
+        //    if (accessor != null)
+        //    {
+        //        double value = accessor();
+        //        if (double.IsNaN(value))
+        //        {
+        //            accessor = null;
+        //        }
+        //    }
 
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateDragForce(); };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
-            }
-        }
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateDragForce(); };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
+        //    }
+        //}
 
-        private VariableEvaluator DynamicPressure()
-        {
-            Func<double> accessor = null;
+        //private VariableEvaluator DynamicPressure()
+        //{
+        //    Func<double> accessor = null;
 
-            accessor = (Func<double>)GetInternalMethod("JSIFAR:GetDynamicPressure", typeof(Func<double>));
-            if (accessor != null)
-            {
-                double value = accessor();
-                if (double.IsNaN(value))
-                {
-                    accessor = null;
-                }
-            }
+        //    accessor = (Func<double>)GetInternalMethod("JSIFAR:GetDynamicPressure", typeof(Func<double>));
+        //    if (accessor != null)
+        //    {
+        //        double value = accessor();
+        //        if (double.IsNaN(value))
+        //        {
+        //            accessor = null;
+        //        }
+        //    }
 
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return vessel.dynamicPressurekPa; };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
-            }
-        }
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return vessel.dynamicPressurekPa; };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
+        //    }
+        //}
 
-        private VariableEvaluator LandingError()
-        {
-            Func<double> accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingError", typeof(Func<double>));
+        //private VariableEvaluator LandingError()
+        //{
+        //    Func<double> accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingError", typeof(Func<double>));
 
-            return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
-        }
+        //    return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
+        //}
 
-        private VariableEvaluator LandingAltitude()
-        {
-            Func<double> accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingAltitude", typeof(Func<double>));
+        //private VariableEvaluator LandingAltitude()
+        //{
+        //    Func<double> accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingAltitude", typeof(Func<double>));
 
-            return (string variable, RasterPropMonitorComputer rpmComp) => 
-            {
-                double est = accessor();
-                return (est == 0.0) ? estLandingAltitude : est;
-            };
-        }
+        //    return (string variable, RasterPropMonitorComputer rpmComp) => 
+        //    {
+        //        double est = accessor();
+        //        return (est == 0.0) ? estLandingAltitude : est;
+        //    };
+        //}
 
-        private VariableEvaluator LandingLatitude()
-        {
-            Func<double> accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingLatitude", typeof(Func<double>));
+        //private VariableEvaluator LandingLatitude()
+        //{
+        //    Func<double> accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingLatitude", typeof(Func<double>));
 
-            return (string variable, RasterPropMonitorComputer rpmComp) =>
-            {
-                double est = accessor();
-                return (est == 0.0) ? estLandingLatitude : est;
-            };
-        }
+        //    return (string variable, RasterPropMonitorComputer rpmComp) =>
+        //    {
+        //        double est = accessor();
+        //        return (est == 0.0) ? estLandingLatitude : est;
+        //    };
+        //}
 
-        private VariableEvaluator LandingLongitude()
-        {
-            Func<double> accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingLongitude", typeof(Func<double>));
+        //private VariableEvaluator LandingLongitude()
+        //{
+        //    Func<double> accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingLongitude", typeof(Func<double>));
 
-            return (string variable, RasterPropMonitorComputer rpmComp) =>
-            {
-                double est = accessor();
-                return (est == 0.0) ? estLandingLongitude : est;
-            };
-        }
+        //    return (string variable, RasterPropMonitorComputer rpmComp) =>
+        //    {
+        //        double est = accessor();
+        //        return (est == 0.0) ? estLandingLongitude : est;
+        //    };
+        //}
 
-        private VariableEvaluator LiftAccel()
-        {
-            Func<double> accessor = null;
+        //private VariableEvaluator LiftAccel()
+        //{
+        //    Func<double> accessor = null;
 
-            accessor = (Func<double>)GetInternalMethod("JSIFAR:GetLiftForce", typeof(Func<double>));
-            if (accessor != null)
-            {
-                double value = accessor();
-                if (double.IsNaN(value))
-                {
-                    accessor = null;
-                }
-            }
+        //    accessor = (Func<double>)GetInternalMethod("JSIFAR:GetLiftForce", typeof(Func<double>));
+        //    if (accessor != null)
+        //    {
+        //        double value = accessor();
+        //        if (double.IsNaN(value))
+        //        {
+        //            accessor = null;
+        //        }
+        //    }
 
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateLiftForce() / totalShipWetMass; };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor() / totalShipWetMass; };
-            }
-        }
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateLiftForce() / totalShipWetMass; };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor() / totalShipWetMass; };
+        //    }
+        //}
 
-        private VariableEvaluator LiftForce()
-        {
-            Func<double> accessor = null;
+        //private VariableEvaluator LiftForce()
+        //{
+        //    Func<double> accessor = null;
 
-            accessor = (Func<double>)GetInternalMethod("JSIFAR:GetLiftForce", typeof(Func<double>));
-            if (accessor != null)
-            {
-                double value = accessor();
-                if (double.IsNaN(value))
-                {
-                    accessor = null;
-                }
-            }
+        //    accessor = (Func<double>)GetInternalMethod("JSIFAR:GetLiftForce", typeof(Func<double>));
+        //    if (accessor != null)
+        //    {
+        //        double value = accessor();
+        //        if (double.IsNaN(value))
+        //        {
+        //            accessor = null;
+        //        }
+        //    }
 
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateLiftForce(); };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
-            }
-        }
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateLiftForce(); };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
+        //    }
+        //}
 
-        private VariableEvaluator MechJebAvailable()
-        {
-            Func<bool> accessor = null;
+        //private VariableEvaluator MechJebAvailable()
+        //{
+        //    Func<bool> accessor = null;
 
-            accessor = (Func<bool>)GetInternalMethod("JSIMechJeb:GetMechJebAvailable", typeof(Func<bool>));
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return false; };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor().GetHashCode(); };
-            }
-        }
+        //    accessor = (Func<bool>)GetInternalMethod("JSIMechJeb:GetMechJebAvailable", typeof(Func<bool>));
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return false; };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor().GetHashCode(); };
+        //    }
+        //}
 
-        private float EvaluateSideSlip()
-        {
-            if (evaluateSideSlip == null)
-            {
-                Func<double> accessor = null;
+        //private float EvaluateSideSlip()
+        //{
+        //    if (evaluateSideSlip == null)
+        //    {
+        //        Func<double> accessor = null;
 
-                accessor = (Func<double>)GetInternalMethod("JSIFAR:GetSideSlip", typeof(Func<double>));
-                if (accessor != null)
-                {
-                    double value = accessor();
-                    if (double.IsNaN(value))
-                    {
-                        accessor = null;
-                    }
-                }
+        //        accessor = (Func<double>)GetInternalMethod("JSIFAR:GetSideSlip", typeof(Func<double>));
+        //        if (accessor != null)
+        //        {
+        //            double value = accessor();
+        //            if (double.IsNaN(value))
+        //            {
+        //                accessor = null;
+        //            }
+        //        }
 
-                if (accessor == null)
-                {
-                    evaluateSideSlip = FallbackEvaluateSideSlip;
-                }
-                else
-                {
-                    evaluateSideSlip = accessor;
-                }
-            }
+        //        if (accessor == null)
+        //        {
+        //            evaluateSideSlip = FallbackEvaluateSideSlip;
+        //        }
+        //        else
+        //        {
+        //            evaluateSideSlip = accessor;
+        //        }
+        //    }
 
-            return (float)evaluateSideSlip();
-        }
+        //    return (float)evaluateSideSlip();
+        //}
 
-        private VariableEvaluator SideSlip()
-        {
-            Func<double> accessor = null;
+        //private VariableEvaluator SideSlip()
+        //{
+        //    Func<double> accessor = null;
 
-            accessor = (Func<double>)GetInternalMethod("JSIFAR:GetSideSlip", typeof(Func<double>));
-            if (accessor != null)
-            {
-                double value = accessor();
-                if (double.IsNaN(value))
-                {
-                    accessor = null;
-                }
-            }
+        //    accessor = (Func<double>)GetInternalMethod("JSIFAR:GetSideSlip", typeof(Func<double>));
+        //    if (accessor != null)
+        //    {
+        //        double value = accessor();
+        //        if (double.IsNaN(value))
+        //        {
+        //            accessor = null;
+        //        }
+        //    }
 
-            if (accessor == null)
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateSideSlip(); };
-            }
-            else
-            {
-                return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
-            }
-        }
+        //    if (accessor == null)
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return FallbackEvaluateSideSlip(); };
+        //    }
+        //    else
+        //    {
+        //        return (string variable, RasterPropMonitorComputer rpmComp) => { return accessor(); };
+        //    }
+        //}
 
-        private double TerminalVelocity()
-        {
-            if (evaluateTerminalVelocity == null)
-            {
-                Func<double> accessor = null;
+        //private double TerminalVelocity()
+        //{
+        //    if (evaluateTerminalVelocity == null)
+        //    {
+        //        Func<double> accessor = null;
 
-                accessor = (Func<double>)GetInternalMethod("JSIFAR:GetTerminalVelocity", typeof(Func<double>));
-                if (accessor != null)
-                {
-                    double value = accessor();
-                    if (value < 0.0)
-                    {
-                        accessor = null;
-                    }
-                }
+        //        accessor = (Func<double>)GetInternalMethod("JSIFAR:GetTerminalVelocity", typeof(Func<double>));
+        //        if (accessor != null)
+        //        {
+        //            double value = accessor();
+        //            if (value < 0.0)
+        //            {
+        //                accessor = null;
+        //            }
+        //        }
 
-                if (accessor == null)
-                {
-                    accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetTerminalVelocity", typeof(Func<double>));
-                    double value = accessor();
-                    if (double.IsNaN(value))
-                    {
-                        accessor = null;
-                    }
-                }
+        //        if (accessor == null)
+        //        {
+        //            accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetTerminalVelocity", typeof(Func<double>));
+        //            double value = accessor();
+        //            if (double.IsNaN(value))
+        //            {
+        //                accessor = null;
+        //            }
+        //        }
 
-                if (accessor == null)
-                {
-                    accessor = FallbackEvaluateTerminalVelocity;
-                }
+        //        if (accessor == null)
+        //        {
+        //            accessor = FallbackEvaluateTerminalVelocity;
+        //        }
 
-                evaluateTerminalVelocity = accessor;
-            }
+        //        evaluateTerminalVelocity = accessor;
+        //    }
 
-            return evaluateTerminalVelocity();
-        }
+        //    return evaluateTerminalVelocity();
+        //}
 
-        private double TimeToImpact()
-        {
-            if (evaluateTimeToImpact == null)
-            {
-                Func<double> accessor = null;
+        //private double TimeToImpact()
+        //{
+        //    if (evaluateTimeToImpact == null)
+        //    {
+        //        Func<double> accessor = null;
 
-                if (accessor == null)
-                {
-                    accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingTime", typeof(Func<double>));
-                    double value = accessor();
-                    if (double.IsNaN(value))
-                    {
-                        accessor = null;
-                    }
-                }
+        //        if (accessor == null)
+        //        {
+        //            accessor = (Func<double>)GetInternalMethod("JSIMechJeb:GetLandingTime", typeof(Func<double>));
+        //            double value = accessor();
+        //            if (double.IsNaN(value))
+        //            {
+        //                accessor = null;
+        //            }
+        //        }
 
-                if (accessor == null)
-                {
-                    accessor = FallbackEvaluateTimeToImpact;
-                }
+        //        if (accessor == null)
+        //        {
+        //            accessor = FallbackEvaluateTimeToImpact;
+        //        }
 
-                evaluateTimeToImpact = accessor;
-            }
+        //        evaluateTimeToImpact = accessor;
+        //    }
 
-            double timeToImpact = evaluateTimeToImpact();
-            if (double.IsNaN(timeToImpact) || timeToImpact > 365.0 * 24.0 * 60.0 * 60.0 || timeToImpact < 0.0)
-            {
-                timeToImpact = -1.0;
-            }
-            else if(timeToImpact == 0.0)
-            {
-                return estLandingUT - Planetarium.GetUniversalTime();
-            }
-            return timeToImpact;
-        }
-        #endregion
+        //    double timeToImpact = evaluateTimeToImpact();
+        //    if (double.IsNaN(timeToImpact) || timeToImpact > 365.0 * 24.0 * 60.0 * 60.0 || timeToImpact < 0.0)
+        //    {
+        //        timeToImpact = -1.0;
+        //    }
+        //    else if(timeToImpact == 0.0)
+        //    {
+        //        return estLandingUT - Planetarium.GetUniversalTime();
+        //    }
+        //    return timeToImpact;
+        //}
+        //#endregion
 
         //--- Fallback evaluators
         #region FallbackEvaluators
-        private double FallbackEvaluateAngleOfAttack()
+        internal double FallbackEvaluateAngleOfAttack()
         {
             // Code courtesy FAR.
             Transform refTransform = vessel.GetTransform();
@@ -464,7 +455,7 @@ namespace JSI
             return (double)AoA;
         }
 
-        private double FallbackEvaluateDragForce()
+        internal double FallbackEvaluateDragForce()
         {
             // Equations based on https://github.com/NathanKell/AeroGUI/blob/master/AeroGUI/AeroGUI.cs and MechJeb.
             double dragForce = 0.0;
@@ -510,7 +501,7 @@ namespace JSI
             return dragForce;
         }
 
-        private double FallbackEvaluateSideSlip()
+        internal double FallbackEvaluateSideSlip()
         {
             // Code courtesy FAR.
             Transform refTransform = vessel.GetTransform();
@@ -527,7 +518,7 @@ namespace JSI
             return (double)sideslipAngle;
         }
 
-        private double FallbackEvaluateLiftForce()
+        internal double FallbackEvaluateLiftForce()
         {
             double liftForce = 0.0;
 
@@ -571,7 +562,7 @@ namespace JSI
             return liftForce;
         }
 
-        private double FallbackEvaluateTerminalVelocity()
+        internal double FallbackEvaluateTerminalVelocity()
         {
             // Terminal velocity computation based on MechJeb 2.5.1 or one of the later snapshots
             if (altitudeASL > vessel.mainBody.RealMaxAtmosphereAltitude())
@@ -628,7 +619,7 @@ namespace JSI
         /// since precise is also computationally expensive.
         /// </summary>
         /// <returns></returns>
-        private double FallbackEvaluateTimeToImpact()
+        internal double FallbackEvaluateTimeToImpact()
         {
             double secondsToImpact;
             if (vessel.situation == Vessel.Situations.SUB_ORBITAL || vessel.situation == Vessel.Situations.FLYING)
