+++ RasterPropMonitor/RasterPropMonitor.csproj
@@ -79,7 +79,7 @@
     <Compile Include="Core\RPMCPersistence.cs" />
     <Compile Include="Core\RPMGlobals.cs" />
     <Compile Include="Core\RPMVCEvaluators.cs" />
-    <Compile Include="Core\RPMVCTriggeredEvent.cs" />
+    <Compile Include="Core\RPMCTriggeredEvent.cs" />
     <Compile Include="Core\RPMVCVariableToObject.cs" />
     <Compile Include="Core\RPMVesselComputer.cs" />
     <Compile Include="Core\MappedVariable.cs" />
